requirements.txt
pymupdf
numpy
scikit-learn
google-generativeai
pymongo
sentence-transformers
python-dotenv